+++ src/meson.build
@@ -49,14 +49,7 @@ subdir('system')
 
 subdir('geometry')
 
-window = py.extension_module(
-    'window',
-    'window.c',
-    c_args: warnings_error,
-    dependencies: pg_base_deps,
-    install: true,
-    subdir: pg,
-)
+subdir('window')
 
 # TODO: support SDL3
 if sdl_api != 3
