+++ src/window/window.c
@@ -6,7 +6,7 @@
 #include "pgopengl.h"
 
 #include "doc/sdl2_video_doc.h"
-#include "doc/window_doc.h"
+#include "window_doc.h"
 
 static int is_window_mod_init = 0;
